upload

+++ modules/commands_fn.py
@@ -2,10 +2,18 @@ import discord
 import requests
 from discord.ext import commands
 import time
+import asyncio
+import fortnite_api
+import feedparser
+from .bot import PagerWithEmojis
+
+api = fortnite_api.FortniteAPI(api_key="")
+
 
 class Commands_fn(commands.Cog, name='fortnite'):
 	def __init__(self, bot):
 		self.bot = bot
+
 	@commands.command()
 	async def map(self, ctx):
 		"""fortnite mapを表示する"""
@@ -31,7 +39,7 @@ class Commands_fn(commands.Cog, name='fortnite'):
 	@commands.command()
 	async def fn(self, ctx, *args):
 		"""Fortnite 成績を表示する fn <ユーザー名>"""
-		await ctx.message.delete()
+		#await ctx.message.delete()
 		joinedArgs = ('+'.join(args))
 		edit = await ctx.send(f'{joinedArgs}のデータを取得中……')
 		res_lang = "ja"
@@ -60,11 +68,52 @@ class Commands_fn(commands.Cog, name='fortnite'):
 			embed.set_footer(text=f'送信者:{ctx.author.display_name}')
 			await edit.edit(content="", embed=embed)
 		if ctx.channel.id == 794100709271535646:
-		  send = True
-		  if send :
-		    time.sleep(2)
-		    embed=discord.Embed(title="使い方", description="```.fn <ユーザー名>```Fortniteの成績を表示します", color=0x4273b5)
-		    embed.set_footer(text=".helpでBOTのヘルプを表示")
-		    await ctx.send(embed=embed)
-		  time.sleep(20)
-		  await edit.delete()
+			if False:
+				time.sleep(2)
+				embed = discord.Embed(
+				    title="使い方",
+				    description="```.fn <ユーザー名>```Fortniteの成績を表示します",
+				    color=0x4273b5)
+				embed.set_footer(text=".helpでBOTのヘルプを表示")
+				await ctx.send(embed=embed)
+				await asyncio.sleep(20)
+				await edit.delete()
+
+	@commands.command()
+	async def fnjpnews(self, ctx, *args):
+		RSS_URL = 'https://fnjpnews.com/feed'
+		d = feedparser.parse(RSS_URL)
+		data = ""
+		for entry in d.entries:
+			data = data + f'[{entry.title}]({entry.link})\n'
+			await asyncio.sleep(1)
+			embed = discord.Embed(title=f'{entry.title}', color=0x00ff00)
+			embed.add_field(
+			    name="description",
+			    value=f'{entry.description}\n\n[リンク]({entry.link})')
+			embed.set_footer(text=f'送信者:{ctx.author.display_name}')
+			message = await ctx.send(embed=embed)
+		embed = discord.Embed(title=f'fnjpnews.com', color=0x00ff00)
+		embed.add_field(name="一覧" ,value=data)
+		await ctx.send(embed=embed)
+	@commands.command()
+	async def page_test(ctx: commands.Context):
+	    pages: list[discord.Embed] = [
+	        discord.Embed(
+	            title="1ページ目です。",
+	            description="最初のページなので、右矢印とバツマークのリアクションが追加されます。",
+	            color=discord.Color.random()
+	        ),
+	        discord.Embed(
+	            title="2ページ目です。",
+	            description="左矢印と右矢印に、バツマークのリアクションも追加されます。",
+	            color=discord.Color.random()
+	        ),
+	        discord.Embed(
+	            title="最後のページです。",
+	            description="最後のページなので、左矢印とバツマークのリアクションが追加されます。",
+	            color=discord.Color.random()
+	        )
+	    ]
+	    pager = PagerWithEmojis(pages)
+	    await pager.discord_pager(ctx)
